Add logic for image reordering in Delete endpoint

+++ Catalog/Catalog.Api/Controllers/v1/ProductController.cs
@@ -385,9 +385,19 @@ public class ProductController : ApiControllerBase<ProductController>, IProductC
         if (productImageEntity.ProductId != productId)
             return BadRequest("The product ID provided does not match the product ID associated with the image.");
 
-        var isRemoved = await productImageRepository.RemoveByIdAsync(productImageId);
+        var allImages = await productImageRepository.GetAllByProductIdAsync(productImageEntity.ProductId);
+        allImages.RemoveAll(x => x.Id == productImageEntity.Id);
 
-        if (isRemoved)
+        var orderedImages = allImages.OrderBy(x => x.DisplayOrder).ToList();
+        for (var i = 0; i < orderedImages.Count; i++)
+        {
+            orderedImages[i].Update(i);
+        }
+
+        var isRemoved = await productImageRepository.RemoveByIdAsync(productImageId);
+        var isUpdated = await productImageRepository.BatchUpdateAsync(orderedImages);
+
+        if (isRemoved && isUpdated)
         {
             await blobService.DeleteFileAsync(blobServiceSettings.ProductImageBucketName, productImageEntity.ImageFileName);
